+++ src/main/kotlin/step3_4/validator/InputStringValidator.kt
@@ -1,4 +1,4 @@
-package step3.validator
+package step3_4.validator
 
 object InputStringValidator {
     /**
